Separate OAuth methods in separate file
